move much of the galsim image making machinery to a common module.
